+++ src/lua/luna.hpp
@@ -177,7 +177,7 @@ namespace lua
 
                     if (lua_isnumber(L, -1)) { /* Check if we got a valid index */
 
-                        int _index = lua_tonumber(L, -1);
+                        int _index = (int)lua_tonumber(L, -1);
 
                         T** obj = static_cast<T**>(lua_touserdata(L, 1));
 
@@ -216,7 +216,7 @@ namespace lua
                     if ( lua_isnumber(L, -1) ) /* Check if we got a valid index */
                     {
 
-                        int _index = lua_tonumber(L, -1);
+                        int _index = (int)lua_tonumber(L, -1);
 
                         T** obj = static_cast<T**>(lua_touserdata(L, 1));
